# Conflicts:
#	src/refresh/model_iqm.c
#	src/refresh/vkpt/bsp_mesh.c
#	src/refresh/vkpt/main.c
#	src/refresh/vkpt/material.c

+++ inc/refresh/models.h
@@ -88,7 +88,7 @@ typedef struct
 	float* tangents;
 	byte* colors;
     byte* blend_indices; // byte4 per vertex
-	float* blend_weights; // float4 per vertex
+	byte* blend_weights; // byte4 per vertex
 	
 	joint_name_t* jointNames; // [num_joints * MAX_QPATH]
 	int* jointParents;
